Added the token randomiser
Randomises the players token. Uses colours to denote whether they won or lost. Uses .txt files to store the tokens. You will have to change the directory for your own system. The only place the directory is used is when the variables are first declared near the top of the program. This program is made in Pycharm and I was unable to find a way to clear the screen so it inefficiently prints lots of lines instead of clearing the screen. The program has been entirely commented. The tokens I have used are attached.