Added a logo on start and a play again system
The program will now ask the user if they want to play again at the end of a round. A logo is also printed on start. This users the logo.txt file and also prints in green. Added some error proofing when taking a deposit. Users used to be able to add a negative value to their balance.